// src/shares/enums/share.enums.ts
export enum ShareResourceType {
  DataRoom = 'DATA_ROOM',
  Folder = 'FOLDER',
  File = 'FILE',
}

export enum ShareKind {
  User = 'USER',
  PublicLink = 'PUBLIC_LINK',
}

export enum ShareRole {
  Viewer = 'VIEWER',
}

export const SHARE_RESOURCE_TYPE_ENUM_NAME = 'share_resource_type_enum';
export const SHARE_KIND_ENUM_NAME = 'share_kind_enum';
export const SHARE_ROLE_ENUM_NAME = 'share_role_enum';